Resolve deprecation warning in calculate_dewpoint_regression()
Calling groupby with temphum.index.date returns a Series with an index of datetime.date objects instead of a DatetimeIndex, resulting in a "FutureWarning: Indexing a DataFrame with a datetimelike index using a single string to slice the rows, like frame[string], is deprecated and will be removed in a future version. Use frame.loc[string] instead" in the DataFrame creation in the next line.
Hence, use normalize() instead.